% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casti.R
\name{casti}
\alias{casti}
\title{Districts (mestske casti) of Prague and other major cities.}
\format{sf data frame with 142 rows of 4 variables + geometry:}
\source{
ArcČR 500
}
\usage{
casti()
}
\description{
\itemize{
  \item{KOD}{Code of the city part / kod mestske casti}
  \item{NAZEV}{Name of the city part / nazev mestske casti}
  \item{KOD_OBEC}{Code of the city}
  \item{NAZ_OBEC}{Name of the city}
}
}
\examples{
 {
}

}
