% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastat_dataset.R
\name{datastat_dataset}
\alias{datastat_dataset}
\title{Get Dataset from DataStat}
\usage{
datastat_dataset(kod = "NEZ01", naming = "label")
}
\arguments{
\item{kod}{code of the dataset}

\item{naming}{whether to use (human readable) \code{label} or (machine friendly) \code{id}}
}
\value{
data frame containing a DataStat dataset
}
\description{
Experimental function to access the "new" DataStat API of Czech Statistical Office / ČSÚ.
}
\details{
This function takes as argument code of the dataset, and returns its content as a data frame.

The function returns the dataset "as it is", without attempting to translate structure or content from the original Czech language.

To save time & bandwidth attempt is made to cache the dataset, either to tempdir or RCZECHIA_HOME directory, if set.

In case of API failures (CZSO down) the function returns NA, with a message.

Usage of the CZSO DataStat API is governed by CZSO Terms & Conditions -
\url{https://csu.gov.cz/podminky_pro_vyuzivani_a_dalsi_zverejnovani_statistickych_udaju_csu}.
}
\examples{

# Unemployment statistics of Czechia

datastat_dataset("NEZ01")


}
