% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{.downloader}
\alias{.downloader}
\title{Internal function - generic downloader, used to serve the rds files from S3}
\usage{
.downloader(file)
}
\arguments{
\item{file}{file to be downloaded (or not...) from S3}
}
\description{
Internal function - generic downloader, used to serve the rds files from S3
}
\keyword{internal}
