% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.rhome_state}
\alias{.rhome_state}
\title{Internal function - executes on load, and informs about state of the RCZECHIA_HOME environment variable}
\usage{
.rhome_state()
}
\description{
The function takes no parameters; it serves the need to make the .onAttach code testable
}
\keyword{internal}
