% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctverce.R
\name{KFME_grid}
\alias{KFME_grid}
\title{Grids cells (faunistické čtverce) of the Czech Republic}
\format{\code{sf} data frame with 1092 rows

\describe{
  \item{ctverec}{KFME code of the grid cell}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
KFME_grid()
}
\description{
Function returning grid covering the Czech Republic according to the Kartierung der Flora Mitteleuropas methodology.
}
\examples{
library(ggplot2)

ggplot() +
   geom_sf(data = republika("low")) +
   geom_sf(data = KFME_grid(), fill = NA)

}
