% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesy.R
\name{lesy}
\alias{lesy}
\title{Woodland Areas}
\format{
\code{sf} data frame with 2.366 rows of geometry variable only
}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
lesy()
}
\description{
Function returning data frame of woodland areas (lesy) of more than 30 hectares in are of the Czech Republic as \code{sf} polygons. It has no obligatory parameters.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).
The data is current to September 2016. Downloaded size is 2.1 MB.
}
