% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obce.R
\docType{data}
\name{obce}
\alias{obce}
\title{Inhabited places / communes (obce) of the Czech Republic.}
\format{6.258 communes (obcí) of the Czech Republic, as a SpatialPolygonsDataFrame:
\describe{
  \item{KOD_OBEC}{Code of the commune (obec), primary key. Use this as key to add other data items.}
  \item{KOD_LAU1}{Code of the district (okres).}
  \item{KOD_CZNUTS3}{Code of the region (kraj).}
  \item{Obec}{Name of the commune. With Czech accents.}
  \item{KOD_POV}{Code of the level II commune (dvojková obec - obec s pověřeným obecním úřadem).}
  \item{PovObec}{Name of the level II commune (dvojková obec - obec s pověřeným obecním úřadem).}
  \item{KOD_ROZ}{Code of the level III commune (trojková obec - obec s rozšířenou působností).}
  \item{RozObec}{Name of the level III commune (trojková obec - obec s rozšířenou působností).}
  \item{Okres}{Name of the district (okres). With Czech accents.}
  \item{Kraj}{Name of the region (kraj). With Czech accents.}
  \item{Obyvatel}{Population, as at 2017-01-01. Source: Czech Statistial Office}
  \item{Obyvatel15p}{Population 15+, as at 2017-01-01. Source: Czech Statistial Office}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
obce
}
\description{
A Large SpatialPolygonsDataFrame containing the communes (obce) of the Czech Republic.
}
\keyword{datasets}
