% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok_to_proceed.R
\name{ok_to_proceed}
\alias{ok_to_proceed}
\title{Internal function - tests availabilty of interet resources}
\usage{
ok_to_proceed(remote_file)
}
\arguments{
\item{remote_file}{resource to be tested}
}
\description{
Internal function - tests availabilty of interet resources
}
\keyword{internal}
