% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresy.R
\docType{data}
\name{okresy}
\alias{okresy}
\title{Districts (okresy) of the Czech Republic. LAU1 administrative unit for the Czech Republic.}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 77 rows and 7 columns.}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
okresy
}
\description{
Districts (okresy) of the Czech Republic. LAU1 administrative unit for the Czech Republic.
}
\keyword{datasets}
