% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orp_polygony.R
\docType{data}
\name{orp_polygony}
\alias{orp_polygony}
\title{Level III Municipalities / communes (obce s rozsirenou pusobnosti) of the Czech Republic}
\format{205 level III municipalities of the Czech Republic + Prague (#3018), as a SpatialPolygonsDataFrame:
\describe{
  \item{KOD_ORP}{Code of the level III commune (obec s rozsirenou pusobnosti).}
  \item{ORP}{Name of the level III commune (obec s rozsirenou pusobnosti). With Czech accents.}
  \item{Obyvatel}{Population, as of 2017-01-01. Source: Czech Statistial Office}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
orp_polygony
}
\description{
A Large SpatialPolygonsDataFrame containing the municipalities of the Czech Republic.
}
\keyword{datasets}
