% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Praha.R
\docType{data}
\name{Praha}
\alias{Praha}
\title{Districts (mestske casti) of Prague.}
\format{57 districts (mestske casti) of Prague, as a sf data frame:
\describe{
  \item{KOD_MOaMC}{Code of the city part / kod mestske casti}
  \item{NAZ_ZKR_MOaMC}{Short name of the city part / zkraceny nazev mestske casti}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
Praha
}
\description{
Districts (mestske casti) of Prague.
}
\keyword{datasets}
