% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCzechia.R
\docType{package}
\name{RCzechia}
\alias{RCzechia}
\alias{RCzechia-package}
\title{RCzechia: package providing spatial objects relevant to the Czech Republic}
\description{
The RCzechia package provides
}
\section{Administrative units}{

*republika* - borders of the Czech Republic
*kraje* - regions / NUTS3 units
*okresy* - districts / LAU1 units
*ORP* - municipalities with extended powers (obce s rozšířenou působností)
*obce_polygony* - municipalities as polygons
*obce_body* - municipalities as centroids (points)
*casti* - city districts (where relevant)
}

\section{Natural objects}{

*reky* - rivers
*plochy* - water bodies
}

