% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reky.R
\name{reky}
\alias{reky}
\title{Rivers}
\format{\code{sf} data frame with 6.198 rows of 4 variables + geometry:

\describe{
  \item{TYP}{Type of river: 1 =  natural, 2 = man-made, 3 = fictional}
  \item{NAZEV}{Name, with Czech accents}
  \item{NAZEV_ASCII}{Name, without Czech accents}
  \item{Major}{Boolean indicating one of the major rivers.}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
reky(method = "curl")
}
\arguments{
\item{method}{Method argument for `download.file()`. The default (i.e. "curl") should be appropriate in most situations.}
}
\description{
Function returning data frame of rivers of the Czech Republic as \code{sf} lines. It takes no parameters.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package). Downloaded size is 1 MB.
}
