% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeo.R
\name{revgeo}
\alias{revgeo}
\title{Reversely Geocode a Czech Address}
\usage{
revgeo(coords)
}
\arguments{
\item{coords}{coordinates to be reverse geocoded; expected as \code{sf} data
frame of spatial points}
}
\value{
\code{sf} data frame as input, with column revgeocoded added (or overwritten)
}
\description{
This function connects to Czech State Administration of Land Surveying and
Cadastre (\url{https://cuzk.gov.cz/en}) API to reversely geocode an address.
As consequence it is implemented only for Czech addresses.
}
\details{
Input of the function is a \code{sf} data frame of spatial points, and
output a vector of characters.

The function returns the same \code{sf} data frame as input, with added field
revgeocoded; it contains the result of operation. Should the data frame contain
a column named revgeocoded it will be overwritten.

In case of reverse geocoding failures (e.g. coordinates outside of the Czech
Republic and therefore scope of ČÚZK) NA is returned.

In case of API failures (CUZK down) the function returns NAs again, with a message.

Usage of the ČÚZK API is governed by ČÚZK Terms & Conditions -
\url{https://geoportal.cuzk.cz/Dokumenty/Podminky.pdf}.
}
\examples{
\donttest{
library(dplyr)
library(sf)

brno <- obce_polygony() |> # shapefile of Brno
  filter(NAZ_OBEC == "Brno") |>
  st_transform(5514) # planar CRS (eastings & northings)

pupek_brna <- st_centroid(brno) # calculate centroid

adresa_pupku <- revgeo(pupek_brna)$revgeocoded # address of the center

print(adresa_pupku)
}
}
