% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_home}
\alias{set_home}
\title{Set the local cache directory}
\usage{
set_home(path)
}
\arguments{
\item{path}{file to be downloaded (or not...) from S3}
}
\value{
TRUE for success and FALSE for failure; returned silently
}
\description{
The function sets the environment variable RCZECHIA_HOME to be used as a local cache for RCzechia remote files; if unset tempdir() is used instead, with persistence for current session only.
}
\details{
Note that when set (it is unset by default) the remote files will be cached to local file system, for good or bad. Also note that you can set the value of RCZECHIA_HOME environment variable directly, either via \code{Sys.setenv()} or via your .Rprofile.
}
