% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnice.R
\name{silnice}
\alias{silnice}
\title{Road Network}
\source{
Mapový podklad – Data200, 2021 © Český úřad zeměměřický a katastrální. \url{https://cuzk.gov.cz}
}
\usage{
silnice()
}
\value{
\code{sf} data frame with 59.594 rows of 3 variables + geometry:

\describe{
\item{TRIDA}{Class of the road:}
\itemize{
\item highway = dálnice,
\item speedway = rychlostní silnice,
\item 1st class road = silnice I. třídy,
\item 2nd class road = silnice II. třídy,
\item 3rd class road = silnice III. třídy,
\item other road = neevidovaná silnice
}
\item{CISLO_SILNICE}{Local road code}
\item{MEZINARODNI_OZNACENI}{International road code}
}
}
\description{
Function returning data frame of roads of the Czech Republic as \code{sf} lines. It has no obligatory parameters.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).

The data is current to December 2020. Downloaded size is 6 MB.
}
