% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionSF.R
\name{unionSF}
\alias{unionSF}
\title{Aggregate Polygons in a sf object}
\usage{
unionSF(.data, key)
}
\arguments{
\item{.data}{sf data frame to be aggregated}

\item{key}{name of column to define the output objects}
}
\description{
The function aggregates polygons of a sf data frame according to values of a single column. It has outcome comparable to `unionSpatialPolygons` from `maptools` package, except that it works for sf and not sp objects.
}
\details{
The function has data frame on the first position, so it is pipe friendly. It retains only geometry and key value, dropping all other columns.
}
\examples{
library(sf)

kraje <- unionSF(okresy(), "KOD_CZNUTS3")
  # returns CZ NUTS3 regions of Czech Republic = equivalent to kraje() in geometry, with a bunch of columns missing


}
