% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unset_home}
\alias{unset_home}
\title{Unset the local cache directory}
\usage{
unset_home()
}
\value{
TRUE for success and FALSE for failure; returned silently
}
\description{
The function unsets the environment variable RCZECHIA_HOME, meaning tempdir() will be used in future function calls, and no persistent data will be stored locally.
}
