% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casti.R
\docType{data}
\name{casti}
\alias{casti}
\title{Districts (mestske casti) of Prague and other major cities.}
\format{57 districts (mestske casti) of Prague, as a sf data frame:
\describe{
  \item{KOD}{Code of the city part / kod mestske casti}
  \item{NAZEV}{Name of the city part / nazev mestske casti}
  \item{KOD_OBEC}{Code of the city}
  \item{NAZ_OBEC}{Name of the city}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
casti
}
\description{
Districts (mestske casti) of Prague and other major cities.
}
\keyword{datasets}
