% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casti.R
\name{casti}
\alias{casti}
\title{City Parts}
\format{\code{sf} data frame with 142 rows of 4 variables + geometry

\describe{
  \item{KOD}{Code of the city part / kod mestske casti}
  \item{NAZEV}{Name of the city part / nazev mestske casti}
  \item{KOD_OBEC}{Code of the city}
  \item{NAZ_OBEC}{Name of the city}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
casti()
}
\description{
Function taking no parameters and returning data frame of districts of Prague and other major cities as \code{sf} polygons.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package). Downloaded size is 593.6 KB.
}
