% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_uzemi.R
\name{chr_uzemi}
\alias{chr_uzemi}
\title{Protected Natural Areas}
\format{\code{sf} data frame with 36 rows of 5 variables + geometry

\describe{
  \item{TYP}{Type of protected area: Národní park, Chráněná krajinná oblast}
  \item{NAZEV}{Name, with Czech accents}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
chr_uzemi()
}
\description{
Function returning data frame of protected natural areas (Chráněná území) of the Czech Republic as \code{sf} polygons. It has no obligatory parameters.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package). Downloaded size is 114 KB.
}
