% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastat_catalogue.R
\name{datastat_catalogue}
\alias{datastat_catalogue}
\title{Get Catalogue of DataStat Reports}
\usage{
datastat_catalogue()
}
\value{
data frame with codes and descriptions of DataStat datasets
}
\description{
Experimental function to access the "new" DataStat API of Czech Statistical Office / ČSÚ.
}
\details{
This function takes no arguments, and returns a data frame with data catalogue of DataStat data sets, the most important part of which is the code (kod) and description (nazev).

The function returns the dataset "as it is", without attempting to translate structure or content from the original Czech language.

In case of API failures (CZSO down) the function returns NA, with a message.

Usage of the CZSO DataStat API is governed by CZSO Terms & Conditions -
\url{https://csu.gov.cz/podminky_pro_vyuzivani_a_dalsi_zverejnovani_statistickych_udaju_csu}.
}
\examples{
# DataStat datasets related to unemployment ("nezaměstnanost" in Czech)

datastat_catalogue() |>
  subset(grepl("nezam", nazev))


}
