% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastat_dataset.R
\name{datastat_dataset}
\alias{datastat_dataset}
\title{Get Dataset from DataStat}
\usage{
datastat_dataset(kod = "NEZ01", naming = "label")
}
\arguments{
\item{kod}{code of the dataset}

\item{naming}{whether to use (human readable) \code{label} or (machine friendly) \code{id}}
}
\value{
data frame containing a DataStat dataset
}
\description{
Experimental function to access the "new" DataStat API of Czech Statistical Office / ČSÚ.
}
\details{
This function takes as argument code of the dataset, and returns its content as a data frame.

The function leverages the functionality of JSON_STAT data format, so that either human readable labels, or machine readable codes are returned - this is driven by \code{naming} parameter to the function, which is passed directly to the \code{naming} parameter of the rjstat::fromJSONstat() function used internally.

The function returns the dataset "as it is", without attempting to translate structure or content from the original Czech language.

In case of API failures (CZSO down) the function returns NA, with a message.

Usage of the CZSO DataStat API is governed by CZSO Terms & Conditions -
\url{https://csu.gov.cz/podminky_pro_vyuzivani_a_dalsi_zverejnovani_statistickych_udaju_csu}.
}
\examples{

# Unemployment statistics of Czechia

unempl <- datastat_dataset("NEZ01", "id") |>
  subset(UZ023H2U == "CZ"  # the country as a whole
         & IndicatorType == "5973DI"  # unemployement rate
         )


plot(value ~ CasR, data = unempl, main = "Unemployment in Czechia",
     ylab = "value", xlab = "year")



}
\seealso{
\code{\link[=datastat_metrics]{datastat_metrics()}} \code{\link[=datastat_dimensions]{datastat_dimensions()}} \code{\link[=datastat_catalogue]{datastat_catalogue()}} \code{\link[rjstat:fromJSONstat]{rjstat::fromJSONstat()}}
}
