% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastat_query.R
\name{datastat_query}
\alias{datastat_query}
\title{Query DataStat Datasets}
\usage{
datastat_query(dimension = NULL, metric = NULL)
}
\arguments{
\item{dimension}{dimension ("dimenze" in Czech) to be queried}

\item{metric}{metric ("ukazatel" in Czech) to be queried}
}
\value{
data frame with codes and descriptions of DataStat datasets
}
\description{
Experimental function to access the "new" DataStat API of Czech Statistical Office / ČSÚ.
}
\details{
This function takes as argument either a dimension or a metric, and returns a data frame listing datasets using the supplied dimension or metric.

The function returns the dataset "as it is", without attempting to translate structure or content from the original Czech language.

In case of API failures (CZSO down) the function returns NA, with a message.

Usage of the CZSO DataStat API is governed by CZSO Terms & Conditions -
\url{https://csu.gov.cz/podminky_pro_vyuzivani_a_dalsi_zverejnovani_statistickych_udaju_csu}.
}
\examples{
\donttest{
# DataStat datasets related to Artificial Intelligence

result <- datastat_query(dimension = "DRUHTECHAI")

result[,c("kod","nazev")]
}

}
