% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{.downloader}
\alias{.downloader}
\title{Internal function - generic downloader, used to serve the rds files from S3}
\usage{
.downloader(file)
}
\arguments{
\item{file}{file to be downloaded (or not...) from S3}
}
\description{
The function utilizes environment variable RCZECHIA_MIRROR as a mirror location of remote files; to configure an alternative (possibly local) repository use \code{Sys.setenv("RCZECHIA_MIRROR" = "a path")}
}
\keyword{internal}
