% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode a Czech Address}
\usage{
geocode(address, crs = 4326)
}
\arguments{
\item{address}{point to be geocoded}

\item{crs}{coordinate reference system of output}
}
\description{
tohle jednou bude vooostrá geokódovací funkce nad RÚIAN
}
