% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode a Czech Address}
\usage{
geocode(address, crs = 4326)
}
\arguments{
\item{address}{point to be geocoded, as character (vector)}

\item{crs}{coordinate reference system of output; default = WGS84}
}
\value{
\code{sf} data frame with 3 variables + geometry

\describe{
\item{address}{the address searched (address input)}
\item{type}{type of record matched by API}
\item{result}{address as returned by API / recorded in RÚIAN}
\item{geometry}{hidden column with spatial point data}
}
}
\description{
This function connects to Czech State Administration of Land Surveying and
Cadastre (\url{https://www.cuzk.cz/en}) API to geocode an address. As
consequence it is implemented only for Czech addresses.
}
\details{
Input of the function are an address to geocode (or a vector of addresses)
and expected Coordinate Reference System of output (default is WGS84 =
EPSG:4326, but in some use cases inž. Křovák = EPSG:5514 may be more
relevant).

NA's in input are considered an error.

Output is a \code{sf} data frame of spatial points.

Depending on the outcome of matching the address to RÚIAN data there is a
number of possible outcomes:

\itemize{
\item{All items were \emph{matched exactly}: the returned \code{sf} data
frame has the same number of rows as there were elements in vector to
be geocoded. The field \emph{target} will have zero duplicates.}
\item{Some items had \emph{multiple matches}: the returned \code{sf} data
frame has more rows than the there were elements in vector to be geocoded.
In the field \emph{target} will be duplicate values. Note that the RÚIAN API
limits multiple matches to 10.}
\item{Some (but not all) items had \emph{no match} in RUIAN data: the returned
\code{sf} data frame will have fewer rows than the vector sent.
to be geocoded elements. Some values will be missing from field \emph{target}}.
\item{No items were matched at all: the function returns empty data frame and a message.}
\item{The CUZK API is down or overloaded: the function returns empty data frame and a message.}
}

Note that character encoding is heavily platform dependent, and you may need to convert to UTF-8,
e.g. by running \code{address <- iconv(address, from = "windows-1250", to = "UTF-8")}
before calling the function.

Usage of the ČÚZK API is governed by ČÚZK Terms & Conditions -
\url{https://geoportal.cuzk.cz/Dokumenty/Podminky.pdf}.
}
\examples{
\donttest{
asdf <- geocode("Gogolova 212, Praha 1")
print(asdf)
}
}
