% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obce_body.R
\docType{data}
\name{obce_body}
\alias{obce_body}
\title{Municipalities / communes (obce) of the Czech Republic. LAU2 administrative unit for the Czech Republic.}
\format{6.258 municipalities of the Czech Republic, as a sf data frame:
\describe{
  \item{KOD_OBEC}{Code of the level I commune (obec).}
<<<<<<< HEAD
  \item{NAZ_OBEC}{Name name of the level I commune (obec).}
=======
  \item{NAZ_OBEC}{Name of the level I commune (obec).}
>>>>>>> 689341d79d655b73a3603192a19deac28436cae4
  \item{KOD_ZUJ}{Code of the basic administrative unit (ICZUJ).}
  \item{NAZ_ZUJ}{Name of the basic administrative unit (ICZUJ).}
  \item{KOD_POU}{Code of the level II commune (obec s poverenym uradem).}
  \item{NAZ_POU}{Name of the level II commune (obec s poverenym uradem)).}
  \item{KOD_ORP}{Code of the level III commune (obec s rozsirenou pusobnosti).}
  \item{NAZ_ORP}{Name of the level III commune (obec s rozsirenou pusobnosti).}
  \item{KOD_OKRES}{Code of the district (okres).}
  \item{KOD_LAU1}{Code of the LAU1 administrative unit (okres).}
  \item{NAZ_LAU1}{Name of the LAU1 administrative unit (okres).}
  \item{KOD_KRAJ}{Code of the region (kraj).}
  \item{KOD_CZNUTS2}{Code of the NUTS3 unit (kraj)}
  \item{NAZ_CZNUTS2}{Name of the NUTS3 unit (kraj)}
}}
\source{
ArcČR 500
}
\usage{
obce_body
}
\description{
Municipalities / communes (obce) of the Czech Republic. LAU2 administrative unit for the Czech Republic.
}
\keyword{datasets}
