% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresy.R
\name{okresy}
\alias{okresy}
\title{Districts (okresy) of the Czech Republic.
LAU1 administrative unit for the Czech Republic.}
\format{sf data frame with 77 rows of 6 variables + geometry:}
\source{
ArcČR 500
}
\usage{
okresy()
}
\description{
\itemize{
  \item{KOD_OKRES}{Code of the district (okres).}
  \item{KOD_LAU1}{Code of the district as LAU1 unit (okres), primary key. Use this as key to add other data items.}
  \item{NAZ_LAU1}{Name of the district as LAU1 unit (okres).}
  \item{KOD_KRAJ}{Code of the region.}
  \item{KOD_CZNUTS3}{Code of the region as NUTS3 (kraj).}
  \item{NAZ_CZNUTS3}{Name of the region (kraj).}
}
}
\examples{
{

}
}
