% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresy.R
\docType{data}
\name{okresy}
\alias{okresy}
\title{Districts (okresy) of the Czech Republic. LAU1 administrative unit for the Czech Republic.}
\format{77 districts of the Czech Republic, as a sf data frame:
\describe{
  \item{KOD_OKRES}{Code of the district (okres).}
  \item{KOD_LAU1}{Code of the district as LAU1 unit (okres), primary key. Use this as key to add other data items.}
  \item{NAZ_LAU1}{Name of the district as LAU1 unit (okres).}
  \item{KOD_KRAJ}{Code of the region.}
  \item{KOD_CZNUTS3}{Code of the region as NUTS3 (kraj).}
  \item{NAZ_CZNUTS3}{Name of the region (kraj).}
}}
\source{
ArcČR 500
}
\usage{
okresy
}
\description{
Districts (okresy) of the Czech Republic. LAU1 administrative unit for the Czech Republic.
}
\keyword{datasets}
