% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orp_polygony.R
\name{orp_polygony}
\alias{orp_polygony}
\title{Municipalities with extended powers (obce s rozšířenou působností) of the Czech Republic}
\format{sf data frame with 205 rows of 10 variables + geometry:}
\source{
ArcČR 500
}
\usage{
orp_polygony()
}
\description{
\itemize{
  \item{KOD_ORP}{Code of the level III commune (obec s rozsirenou pusobnosti).}
  \item{NAZ_ZKR_ORP}{Short name of the level III commune (obec s rozsirenou pusobnosti).}
  \item{NAZ_ORP}{Full name of the level III commune (obec s rozsirenou pusobnosti).}
  \item{KOD_RUIAN}{RUIAN (Registr uzemni identifikace, adres a nemovitosti) code.}
  \item{KOD_OKRES}{Code of the district (okres).}
  \item{KOD_LAU1}{Code of the LAU1 administrative unit (okres).}
  \item{NAZ_LAU1}{Name of the LAU1 administrative unit (okres).}
  \item{KOD_KRAJ}{Code of the region (kraj).}
  \item{KOD_KRAJ}{Code of the region (kraj).}
  \item{KOD_CZNUTS2}{Code of the NUTS3 unit (kraj)}
  \item{NAZ_CZNUTS2}{Name of the NUTS3 unit (kraj)}
}
}
\examples{
{


}
}
