% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orp_polygony.R
\docType{data}
\name{orp_polygony}
\alias{orp_polygony}
\title{Level III Municipalities / communes (obce s rozsirenou pusobnosti) of the Czech Republic}
\format{205 level III municipalities of the Czech Republic , as a sf data frame:
\describe{
  \item{KOD_ORP}{Code of the level III commune (obec s rozsirenou pusobnosti).}
  \item{NAZ_ZKR_ORP}{Short name of the level III commune (obec s rozsirenou pusobnosti).}
  \item{NAZ_ORP}{Full name of the level III commune (obec s rozsirenou pusobnosti).}
  \item{KOD_RUIAN}{RUIAN (Registr uzemní identifikace, adres a nemovitosti) code.}
  \item{KOD_OKRES}{Code of the district (okres).}
  \item{KOD_LAU1}{Code of the LAU1 administrative unit (okres).}
  \item{NAZ_LAU1}{Name of the LAU1 administrative unit (okres).}
  \item{KOD_KRAJ}{Code of the region (kraj).}
  \item{KOD_KRAJ}{Code of the region (kraj).}
  \item{KOD_CZNUTS2}{Code of the NUTS3 unit (kraj)}
  \item{NAZ_CZNUTS2}{Name of the NUTS3 unit (kraj)}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
orp_polygony
}
\description{
Level III Municipalities / communes (obce s rozsirenou pusobnosti) of the Czech Republic
}
\keyword{datasets}
