% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orp_polygony.R
\name{orp_polygony}
\alias{orp_polygony}
\title{Obce s rozsirenou pusobnosti}
\source{
© ČÚZK, 2021 \url{https://vdp.cuzk.cz/}
}
\usage{
orp_polygony()
}
\value{
\code{sf} data frame with 206 rows of 5 variables + geometry

\describe{
\item{KOD_ORP}{Code of the level III commune (obec s rozsirenou pusobnosti).}
\item{NAZ_ORP}{Full name of the level III commune (obec s rozsirenou pusobnosti).}
\item{KOD_KRAJ}{Code of the region (kraj).}
\item{KOD_CZNUTS3}{Code of the NUTS3 unit (kraj)}
\item{NAZ_CZNUTS3}{Name of the NUTS3 unit (kraj)}
}
}
\description{
Function returning data frame of municipalities with extended powers (obce s rozšířenou působností) as \code{sf} polygons. It takes no parameters.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).

The data is current to June 2021 (i.e changes introduced by act 51/2020 Sb. are reflected). Downloaded size is 3.1 MB.
}
