% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plochy.R
\docType{data}
\name{plochy}
\alias{plochy}
\title{Water bodies (vodní plochy) of the Czech Republic}
\format{480 water bodies of the Czech Repubic, as a Large SpatialPolygonsDataFrame.
\describe{
  \item{NAZEV}{Name, in ASCII}
  \item{VYSKA}{water level, meters above sea level}
  \item{Major}{Boolean indicating the major water bodies (= Lipno, Orlík, Slapy, Nechranice, Novomlýnské nádrže.)}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
plochy
}
\description{
A Large SpatialPolygonsDataFrame containing water bodies (vodní plochy) of the Czech Republic.
}
\keyword{datasets}
