% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plochy.R
\name{plochy}
\alias{plochy}
\title{Water Bodies}
\format{
\code{sf} data frame with 28.011 rows of 2 variables + geometry

\describe{
  \item{TYP}{Type of water body}
  \item{NAZEV}{Name, with Czech accents}
}
}
\source{
Mapový podklad – Data50, 2021 © Český úřad zeměměřický a katastrální \url{https://www.cuzk.cz}
}
\usage{
plochy()
}
\description{
Function returning data frame of water bodies of the Czech Republic as \code{sf} polygons. It takes no parameters.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).

The data is current to December 2020. Downloaded size is 15.5 MB.
}
