% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Praha.R
\docType{data}
\name{Praha}
\alias{Praha}
\title{Districts (mestske casti) of Prague.}
\format{57 districts (mestske casti) of Prague:
\describe{
  \item{KOD_MOaMC}{Code of the city part / kod mestske casti}
  \item{NAZ_ZKR_MOaMC}{Short name of the city part / zkraceny nazev mestske casti}
  \item{NAZ_MOaMC}{Full name of the city part / plny nazev mestske casti}
  \item{KOD_ZUJ}{Kod zakladni uzemni jednotky}
  \item{NAZ_ZUJ}{Nazev zakladni uzemni jednotky}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
Praha
}
\description{
A Large SpatialPolygonsDataFrame containing the administrative districts of Prague.
}
\keyword{datasets}
