% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reky.R
\docType{data}
\name{reky}
\alias{reky}
\title{Rivers (reky) of the Czech Republic}
\format{6.198 rivers of the Czech Repubic, as a Large SpatialLinesDataFrame.
\describe{
  \item{NAZEV}{Name, with Czech accents}
  \item{Major}{Boolean indicating the major rivers (= Labe, Vltava, Svratka, Morava, Berounka, Sazava, Odra, Dyje, Mze, Radbuza, Uslava, Ohre, Otava).}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
reky
}
\description{
A Large SpatialLinesDataFrame containing rivers (reky) of the Czech Republic.
}
\keyword{datasets}
