% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reky.R
\docType{data}
\name{reky}
\alias{reky}
\title{Rivers (řeky) of the Czech Republic}
\format{6198 rivers of the Czech Repubic, as Large SpatialLinesDataFrame.
\describe{
  \item{NAZEV}{Name, in ASCII}
  \item{Major}{Boolean indicating the major rivers (Labe, Vltava, Svratka, Morava, Berounka, Sazava, Odra, Dyje, Mze, Radbuza, Uslava, Ohre, Otava).}
}}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
reky
}
\description{
A Large SpatialLinesDataFrame containing rivers (řeky) of the Czech Republic.
}
\keyword{datasets}
