% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reky.R
\name{reky}
\alias{reky}
\title{Rivers}
\format{
\code{sf} data frame with 7.654 rows of 3 variables + geometry:

\describe{
  \item{TYP}{Type of river}
  \item{NAZEV}{Name, with Czech accents}
  \item{Major}{Boolean indicating one of the major rivers.}
}
}
\source{
Mapový podklad – Data50, 2021 © Český úřad zeměměřický a katastrální \url{https://www.cuzk.cz}
}
\usage{
reky(scope = "global")
}
\arguments{
\item{scope}{Should the function return all rivers, or just Vltava in Prague / Svitava & Svratka in Brno?}
}
\description{
Function returning data frame of rivers of the Czech Republic as \code{sf} lines. It takes a single parameter \code{scope} with default "global".
}
\details{
Two special case scopes are defined: Praha (returning the part of Vltava in and around Prague) and Brno (returning Svitava and Svratka near and around Brno).

Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).

The data is current to December 2020. Downloaded size is 33 MB.
}
\examples{
\donttest{
library(sf)

plot(st_geometry(subset(okresy(), KOD_LAU1 == "CZ0642"))) # Brno city
plot(reky("Brno"), add = TRUE) # Svitava & Svratka added to Brno my city plot
}

}
