% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reky.R
\name{reky}
\alias{reky}
\title{Rivers (řeky) of the Czech Republic}
\format{sf data frame with 6.198 rows of 4 variables + geometry:}
\source{
ArcČR 500
}
\usage{
reky()
}
\description{
\itemize{
  \item{TYP}{Type of river: 1 =  natural, 2 = man-made, 3 = fictional}
  \item{NAZEV}{Name, with Czech accents}
  \item{NAZEV_ASCII}{Name, without Czech accents}
  \item{Major}{Boolean indicating one of the major rivers.}
}
}
\examples{
{
}
}
