% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/republika.R
\docType{data}
\name{republika}
\alias{republika}
\title{Boundaries of the Czech Republic.}
\format{A Large SpatialPolygonsDataFrame containing the Czech Republic.}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
republika
}
\description{
A Large SpatialPolygonsDataFrame containing the Czech Republic.
}
\keyword{datasets}
