% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/republika.R
\name{republika}
\alias{republika}
\title{Boundaries of the Czech Republic.}
\format{sf data frame with 1 row of 1 variable + geometry:}
\source{
ArcČR 500
}
\usage{
republika()
}
\description{
Boundaries of the Czech Republic.
}
\examples{
{


}
}
