% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/republika.R
\name{republika}
\alias{republika}
\title{Republika}
\format{
\code{sf} data frame with 1 row of 1 variable + geometry:
}
\source{
© ČÚZK, 2021 \url{https://vdp.cuzk.cz/}
}
\usage{
republika(resolution = "high")
}
\arguments{
\item{resolution}{Should the function return high or low resolution shapefile? Allowed values are "high" (default) and "low". This parameter affects only the geometry column, all other fields remain the same.}
}
\description{
Boundaries of the Czech Republic as \code{sf} polygon.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package).

The data is current to June 2021. Downloaded size of high resolution shapefile is <1 MB.
}
\examples{
\donttest{
library(sf)

hranice <- republika()
plot(hranice, col = "white")
}´
}
