% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/republika.R
\docType{data}
\name{republika}
\alias{republika}
\title{Boundaries of the Czech Republic.}
\format{Boundaries of the Czech Republic, as a sf data frame:}
\source{
ArcČR 500
}
\usage{
republika
}
\description{
Boundaries of the Czech Republic.
}
\keyword{datasets}
