% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/republika.R
\name{republika}
\alias{republika}
\title{Republika}
\format{
\code{sf} data frame with 1 row of 1 variable + geometry:
}
\source{
\url{https://www.arcdata.cz/produkty/geograficka-data/arccr-500}
}
\usage{
republika(resolution = "high")
}
\arguments{
\item{resolution}{Should the function return high or low resolution shapefile? Allowed values are "high" (default) and "low". This parameter affects only the geometry column, all other fields remain the same.}
}
\description{
Boundaries of the Czech Republic as \code{sf} polygon.
}
\details{
Due to package size constraints the data are stored externally (and a working internet connection is required to use the package). Downloaded size of high resolution shapefile is 949.7 KB.
}
\examples{
library(sf)

hranice <- republika()
plot(hranice, col = "white")
}
