% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeo.R
\name{revgeo}
\alias{revgeo}
\title{Reversely Geocode a Czech Address}
\usage{
revgeo(coords)
}
\arguments{
\item{coords}{coordinates to be reverse geocoded; expected as sf spatial_points format}
}
\description{
tohle jednou bude vooostrá geokódovací funkce nad RÚIAN
}
